==> T-WAS/Jupyter_notebook/0_preprocessing_case_clean.ipynb <==
#upload some libraries
import os
import pandas.testing as tm
import h5py
import numpy as np
import pandas as pd
import math
import sys
import joblib
import subprocess
import statsmodels.api as sm
from scipy import stats

! pip install --upgrade tables
import tables

# set paths
basedir = '/data/songy4/twas'
datadir = f'{basedir}/data_folder'

Get Pheno (case vs control) data combined with expression data to select case only
##this file should include FID and IID columns at the beginning

#open 10PCs data file to extract pheno information

pheno_df = pd.read_csv(f"{datadir}/baseline_VST_proteincoding_omics_OG_PPMI_AND_PDBP_ALLTRANSCRIPTS_10PCs_loadings.csv", engine='c')

#count how manyID -- total number, both case and control
print("Number of rows in pheno data:", pheno_df.shape[0])

#drop PCs from pheno_df
pheno_id = pheno_df.copy()
pheno_id.drop(pheno_id.iloc[:,2:12], axis=1, inplace=True)
pheno_id.head()

#open normalized expression file PDBP_and_OG_PPMI_Combined_Clinical_Omics.csv

expre_df = pd.read_csv(f"{datadir}/PDBP_and_OG_PPMI_Transcriptomics_ALLTRANSCRIPTS_diffExp_ProteinCoding_ADJUSTED10PCs.csv", engine='c')

#count how manyID -- total number, both case and control
print("Number of rows in expression data:", expre_df.shape[0])

#add PHENO column on omics_df by matching participant_id_x with ID 
expression = expre_df.merge(pheno_id, on='ID', how='left')

#leave only PHENO=1 to only include only cases, not controls
expression_case = expression.loc[expression['PHENO'] == 1.0]

#drop Pheno column
expression_case = expression_case.drop(['PHENO'], axis=1)

#grab ID column and duplicate into two columns to compare with .fam file FID IID
expression_case = expression_case.rename(columns={'ID':'FID'})

#rearrange Gene_Symbol, Chr, Start, Stop columns right nextg to index column 
expression_case.insert(1, 'IID', expression_case['FID'])

expression_case = expression_case.reset_index(drop=True)
print('shape of case only expression data:', expression_case.shape)
expression_case.head()

#case only expression dataset is ready to use

#Example_phenotype_file.txt from github
example = 'https://raw.githubusercontent.com/opain/Calculating-FUSION-TWAS-weights-pipeline/master/Example_phenotype_file.txt'
df_example = pd.read_csv(example, sep='\s+')
df_example.head()

#extract genetic PCs from expression data 
!plink --bfile /{datadir}/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI --pca 10 --out /{datadir}/PCA

#read genetic PCs
pcs_df = pd.read_csv(f"{datadir}/PCA.eigenvec", sep=' ', usecols=[1,2,3,4,5,6,7,8,9,10,11], names=['ID','PC1_gen','PC2_gen','PC3_gen','PC4_gen','PC5_gen','PC6_gen','PC7_gen','PC8_gen','PC9_gen','PC10_gen'])
pcs_df

Create fid_df.txt file to remove rows in .bim file if FID is missing

#grab FID and IID to compare with .fam file FID IID
case_id_df= expression_case[['FID', 'IID']].copy()
print('shape of case only ID:', case_id_df.shape)
case_id_df.head()

#save id_df as txt  -- run only once for download
case_id_df.to_csv(r'./data_folder/case_id_df.txt', sep='\t' ,index=False)
#id_df.to_csv(r'./data_folder/fid_df.csv' ,index=False)

Plink file preprocess

#look into genotype data folder to grab PLINK files
! ls /data/LNG/makariousmb/Projects/GenoMLxAMP/clean_wgs/PDBP_AND_OG_PPMI

#copy bed bim fam files into data folder - run only once for download
#these plink files are genotic data (genotype)
! cp -R /data/LNG/makariousmb/Projects/GenoMLxAMP/clean_wgs/PDBP_AND_OG_PPMI/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI.fam /{datadir}/
! cp -R /data/LNG/makariousmb/Projects/GenoMLxAMP/clean_wgs/PDBP_AND_OG_PPMI/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI.bim /{datadir}/
! cp -R /data/LNG/makariousmb/Projects/GenoMLxAMP/clean_wgs/PDBP_AND_OG_PPMI/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI.bed /{datadir}/

#check number of IDs in original .fam file
!cat {datadir}/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI.fam | wc -l

#look at the fam file (fid, iid, mid, pid, gender, affection status)
! head /data/LNG/makariousmb/Projects/GenoMLxAMP/clean_wgs/PDBP_AND_OG_PPMI/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI.fam

#check number of IDs in original .bim file
!cat {datadir}/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI.bim | wc -l

#look at the bim file (variance, chromosome and position, allele1 and allele2)
! head {datadir}/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI.bim 

#filter sample id using fid_df.txt to reduce rows that are not in fid_df.txt
##plink --bfile <genotype file> --keep <id file> --make-bed --out <outfile>
!plink --bfile /{datadir}/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI --keep /{datadir}/case_id_df.txt  --make-bed --out /{datadir}/qc_genotypes_twas

#check the number of IDs in .fam file to make sure it is same numbers as number of rows in exp_df
!cat /{datadir}/qc_genotypes_twas.fam | wc -l

#check number of IDs in original .bim file
!cat {datadir}/qc_genotypes_twas.bim | wc -l

#look at the bim file (variance, chromosome and position, allele1 and allele2)
! head {datadir}/qc_genotypes_twas.bim 

#get frequency from plink file -- original plink files (AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI)
!plink --bfile /{datadir}/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI --bed /{datadir}/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI.bed --freq --out /{datadir}/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI  

#read frq file from original plink files (AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI)
! head {datadir}/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI.frq

#count if MAF is bigger than 0.5
!cat {datadir}/AMP_Euro_sampleQC_variantQC_FINAL_PDBP_AND_OG_PPMI.frq | awk -F"\t" '$5>0.5' | wc -l

#get frequency from plink file -- qc_genotypes_twas
!plink --bfile /{datadir}/qc_genotypes_twas  --bed /{datadir}/qc_genotypes_twas.bed --freq --out /{datadir}/qc_genotypes_twas

#read frq file from qc_genotypes_twas plink files
! head /{datadir}/qc_genotypes_twas.frq

#count if MAF is bigger than 0.5
!cat /{datadir}/qc_genotypes_twas.frq | awk -F"\t" '$5>0.5' | wc -l

#get frequency from plink file -- liftover version
!plink --bfile /{datadir}/qc_genotypes_twas_hg19_lifted  --bed /{datadir}/qc_genotypes_twas_hg19_lifted.bed --freq --out /{datadir}/qc_genotypes_twas_hg19_lifted

#read frq file from qc_genotypes_twas plink files
! head /{datadir}/qc_genotypes_twas_hg19_lifted.frq

#count if MAF is bigger than 0.5 --> this is to check if minor allele is A1
!cat /{datadir}/qc_genotypes_twas_hg19_lifted.frq | awk -F"\t" '$5>0.5' | wc -l

###if MAF value is bigger than 0.5 that is major allele and those alleles (A1 and A2) need to be flipped using plink --reference-allele
### https://zzz.bwh.harvard.edu/plink/reference.shtml

Covariate data preprocess

#open covariate file clinicodemographic_3nov2020.csv
covariate = pd.read_csv(f"{datadir}/clinicodemographic_3nov2020.csv", index_col=False )

print('shape of covariate data:', covariate.shape)
print(covariate.head())

#drop cov_df ID not in exp_case ID
covariate_case = covariate[covariate['ID'].isin(expression_case['FID'])]

print('shape of covariate data:', covariate_case.shape)
print(covariate_case.head())

#merge pcs_df into covariate_df
covar_df_case = covariate_case.merge(pcs_df, on='ID', how='left')

#rename ID to FID
covariate_case = covar_df_case.rename(columns={'ID':'FID'})

#rearrange Gene_Symbol, Chr, Start, Stop columns right nextg to index column 
covariate_case.insert(1, 'IID', covariate_case['FID'])

covariate_case = covariate_case.reset_index(drop=True)
print('shape of covariate data:', covariate_case.shape)

covariate_case.head()

#check unique values in each column
for i in covariate_case.columns[2:8]:
    print('Unique value of ' + i + ':', covariate_case[i].unique())

#check nan numbers in each column 
print('null value in each column: \n', covariate_case.isna().sum())

#check data type in all column
print('data types in each column: \n',covariate_case.dtypes)

#replace NaN with mean for nonbinary and median for binary column
covariate_case['FAMILY_HISTORY'] = covariate_case['FAMILY_HISTORY'].fillna(covariate_case['FAMILY_HISTORY'].median())
covariate_case['UPSIT'] = covariate_case['UPSIT'].fillna(covariate_case['UPSIT'].mean())

#check nan numbers in each column 
print('null value in each column: \n', covariate_case.isna().sum())

#change float to int in column EDUCATION, FAMILY_HISTORY, and UPSIT

##col=(covariate.dtypes=='float64')
col_case=covariate_case.columns[4:7]
covariate_case.loc[:,col_case]=covariate_case.loc[:,col_case].astype(np.int64)

#normalize covariate values for AGE, UPSIT, EDUCATION columns with min_max_scaling

from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
covariate_scale_case = covariate_case.copy()
covariate_scale_case[['AGE', 'UPSIT', 'EDUCATION']] = scaler.fit_transform(covariate_scale_case[['AGE', 'UPSIT', 'EDUCATION']])



# apply the min-max scaling in Pandas using the .min() and .max() methods
#def min_max_scaling(df):
    # copy the dataframe
 #   df_norm = df.copy()
    # apply min-max scaling
  #  for column in df_norm.iloc[:, [2, 4, 6]]:
   #     df_norm[column] = (df_norm[column] - df_norm[column].min()) / (df_norm[column].max() - df_norm[column].min()) 
        
    #return df_norm
    
# call the min_max_scaling function
#covariate_scale_case = min_max_scaling(covariate_case)
print('shape of covariate data:', covariate_scale_case.shape)
covariate_scale_case.head()

#save covariate as txt -- run only once for download
covariate_scale_case.to_csv(r'./data_folder/covariates_case.txt', sep='\t' ,index=False)

Coordinate data preprocess

#open gene_chr_start_stop.tab with column names -- this has ch37 chr and position info
##the file "gene_chr_start_stop.tab" is in GP2_WG/working groups/CD_DAWG/scratch/example_TWAS
chr_pos_df = pd.read_csv(f"{datadir}/gene_chr_start_stop.tab", sep='\t', header=None, index_col=False, names=('ID', 'X.Chr', 'start', 'end'))

#rearrange the columns 
chr_pos_df = chr_pos_df[['X.Chr','start','end','ID']]

print('shape of chromosom position data:', chr_pos_df.shape)
chr_pos_df.head()

#checke chromosome inputs in chr_pos_df
unique_chromosom = chr_pos_df['X.Chr'].unique()
print('unique chromosom:', unique_chromosom)

#checke ID inputs in chr_pos_df -- there is no overlap in ID
nunique_chromosom = chr_pos_df['X.Chr'].nunique()
print('number of unique chromosom:', nunique_chromosom)

#make exp_case column names as list without FID and IID
##this will be all ENSG transcripts numbers in our expression data
trans = expression_case.columns[2:19688].tolist()
len(trans)

#make trans list as a dataframe
fid = pd.DataFrame(trans, columns=["ID"])
print("total number of columns in expression data:", fid.shape[0])

#check how many ENSG are missing from FID that are not in ID
mis_coord = fid[~fid['ID'].isin(chr_pos_df['ID'])]
print("number of columns not in expression data:", mis_coord.shape[0])

#drop rows of chr_pos_df where chr_pos_df ID is not match with exp_case column name list
coord = chr_pos_df[chr_pos_df['ID'].isin(trans)]

#create pre_gene_list as list for swarm job later
pre_gene_list = coord['ID'].tolist()
len(pre_gene_list)

#count how many MT are in coord
mt_coord = coord[coord['X.Chr'].str.contains('MT')]
print("Number of MT chromosom in coordinate data:" , mt_coord['X.Chr'].count())

#count how many X are in coord
y_coord = coord[coord['X.Chr'].str.contains('Y')]
print("Number of Y chromosom in coordinate data:" , y_coord['X.Chr'].count())

#count how many X are in coord
x_coord = coord[coord['X.Chr'].str.contains('X')]
print("Number of X chromosom in coordinate data:" , x_coord['X.Chr'].count())

#count how many X are in coord
gl_coord = coord[coord['X.Chr'].str.contains('GL')]
print("Number of GL chromosom in coordinate data:" , gl_coord['X.Chr'].count())

#drop all the rows contain MT, X, Y, GL in X.Chr 
word_list = ["MT", "X", "Y", "GL"]
coord_df = coord[~coord['X.Chr'].str.contains('|'.join(word_list))]
coord_df = coord_df.reset_index(drop=True)
print("shape of coord_df:", coord_df.shape)
print(coord_df.head())

#check miscoordinate numbers are same 602
no_coord = mis_coord['ID'].tolist()
print("number of missing genes:", len(no_coord))

#read gene_list from liftover coordinate data
gene_id = coord_df['ID']
print("number of genes in gene_id:",len(gene_id))
#create gene_list as list to drop columns in exp_case 
gene_list = coord_df['ID'].tolist()
print("number of genes in gene_list:", len(gene_list))

#save coord_df as txt -- run only once for download
coord_df.to_csv(r'./data_folder/twas_coordinate.txt', sep='\t' ,index=False)

#save gene_list as txt -- run only once for download
gene_id.to_csv(r'./data_folder/gene_list.txt', sep='\t' ,index=False)

#drop columns if not in gene_list_1
final_pheno_case = expression_case.drop(columns=[col for col in expression_case[expression_case.columns[2:19687]] if col not in gene_list])
final_pheno_case

#save expression matrix as txt --run only once
final_pheno_case.to_csv(r'./data_folder/twas_expression_case.txt', sep= '\t', index=False)

plink files (genomic data), expression data(phenomic data), coordinate data, covariate data, and gene list are ready to use

geno_path = f'{datadir}/qc_genotypes_twas'
#plink files to use: {datadir}/qc_genotypes_twas_hg19_lifted
gene_list_path = f'{datadir}/gene_list.txt'
pheno_path = f'{datadir}/expression_matrix_final.txt'
coord_path = f'{datadir}/twas_coordinate.txt'
covar_path = f'{datadir}/covariates.txt'

GWAS summery stat will be preprocessed after 1_liftover process

#open bim file 
bim_df = pd.read_csv(f"{datadir}/qc_genotypes_twas_hg19_lifted.bim", sep='\t', header=None)
bim_df.columns = ['Chr', 'Variant_ID', 'Position', 'Basepair coordinate', 'Allele_1', 'Allele_2']
bim_df.head()

#check sumstat.csv
meta = pd.read_csv("sumstat.csv", sep=' ')
meta.head()

data statistics (to write a paper)

covariates_case = pd.read_csv(f"{datadir}/covariates_case.txt", sep='\t')
covariates_cont = pd.read_csv(f"{datadir}/covariates_control.txt", sep='\t')
covariates_case

#get percentage of MALE, FAMILY_HISTORY, and InfAJ

print('% of MALE column in case (1=MALE, 0=FEMALE): \n', covariates_case['MALE'].value_counts(normalize=True) * 100)
print('% of MALE column in control (1=MALE, 0=FEMALE): \n', covariates_cont['MALE'].value_counts(normalize=True) * 100)

print('% of FAMILY_HISTORY column in case (1=positive, 0=none): \n', covariates_case['FAMILY_HISTORY'].value_counts(normalize=True) * 100)
print('% of FAMILY_HISTORY column in control (1=positive, 0=none): \n', covariates_cont['FAMILY_HISTORY'].value_counts(normalize=True) * 100)

print('% of InfAJ column in case (1=Ashkenazi, 0=no): \n', covariates_case['InfAJ'].value_counts(normalize=True) * 100)
print('% of InfAJ column in control (1=Ashkenazi, 0=no): \n', covariates_cont['InfAJ'].value_counts(normalize=True) * 100)

#get inverse min-max scaling from covariate data and get mean and SD of AGE, UPSIT

covariates_case[['AGE', 'UPSIT', 'EDUCATION']] = scaler.inverse_transform(covariates_case[['AGE', 'UPSIT', 'EDUCATION']])
covariates_cont[['AGE', 'UPSIT', 'EDUCATION']] = scaler.inverse_transform(covariates_cont[['AGE', 'UPSIT', 'EDUCATION']])


print('Mean of AGE column in case: \n', covariates_case['AGE'].mean())
print('Standard Diviation of AGE column in case: \n', covariates_case['AGE'].std())
print('Mean of UPSIT column in case: \n', covariates_case['UPSIT'].mean())
print('Standard Diviation of UPSIT column in case: \n', covariates_case['UPSIT'].std())

print('Mean of AGE column in control: \n', covariates_cont['AGE'].mean())
print('Standard Diviation of AGE column in control: \n', covariates_cont['AGE'].std())
print('Mean of UPSIT column in control: \n', covariates_cont['UPSIT'].mean())
print('Standard Diviation of UPSIT column in control: \n', covariates_cont['UPSIT'].std())
