==> T-WAS/Jupyter_notebook/1_liftover_clean.ipynb <==
import subprocess
import sys
import os
import shutil
import pandas as pd

def shell_do(command, log=False, return_log=False):
    print(f'Executing: {(" ").join(command.split())}', file=sys.stderr)

    res=subprocess.run(command.split(), stdout=subprocess.PIPE)

    if log:
        print(res.stdout.decode('utf-8'))
    if return_log:
        return(res.stdout.decode('utf-8'))

# set paths
basedir = '/data/songy4/twas'
datadir = f'{basedir}/data_folder'
fusiondir = f'{basedir}/fusion_twas'
geno_path = f'{datadir}/qc_genotypes_twas'
gene_list_path = f'{datadir}/gene_list.txt'
pheno_path = f'{datadir}/expression_matrix_final.txt'
coord_path = f'{datadir}/twas_coordinate.txt'
covar_path = f'{datadir}/covariates_case.txt'

# Liftover genotype from hg38 to hg19 to match 1kg ref

bim = pd.read_csv(f'{geno_path}.bim', sep='\t', header=None, names=['chr', 'rsid', 'kb', 'pos', 'a1', 'a2'])
bim

# get chrN:start-end positions for liftover of genotype from hg38 to hg19 to match 1kG LD ref
lift_outname = f'{datadir}/geno_hg38_positions.bed'
bim = pd.read_csv(f'{geno_path}.bim', sep='\t', header=None)
bim.columns = ['chr', 'rsid', 'kb', 'pos', 'a1', 'a2']
bim['chr'] = 'chr' + bim['chr'].astype('str')
bim['end'] = bim['pos'] + 1
lift_out = bim[['chr', 'pos', 'end', 'rsid' ]].copy()
lift_out.to_csv(lift_outname, sep='\t', header=False, index=False)

# pull liftOver from UCSC
#!wget http://hgdownload.soe.ucsc.edu/admin/exe/linux.x86_64/liftOver -P /data/songy4/twas/liftover/
#!wget --timestamping 'ftp://hgdownload.cse.ucsc.edu/goldenPath/hg38/liftOver/hg38ToHg19.over.chain.gz' -O /data/songy4/twas/liftover/hg38ToHg19.over.chain.gz
#!chmod +x /data/songy4/twas/liftover/liftOver

# create command for liftover
liftOver = f'{basedir}/liftover/liftOver'
chainfile = f'{basedir}/liftover/hg38ToHg19.over.chain.gz'
liftover_output = f'{basedir}/geno_hg19_positions.bed'
unlifted = f'{basedir}/unlifted.bed'

liftover_cmd = f'\
{liftOver} {lift_outname} {chainfile} {liftover_output} {unlifted}'

shell_do(liftover_cmd)
liftover_cmd

# create map update file
lifted = pd.read_csv(liftover_output, sep='\t', header=None, names=['chr', 'pos', 'end', 'rsid' ])
print(lifted)
bim = pd.read_csv(f'{geno_path}.bim', sep='\t', header=None, names=['chr', 'rsid', 'kb', 'pos', 'a1', 'a2'])
print(bim)
bim_lift_merge = bim.merge(lifted, how='right', on='rsid')
lifted_bim = bim_lift_merge[['chr_x', 'rsid', 'kb', 'pos_y', 'a1', 'a2']].copy()
lifted_bim.columns = ['chr', 'rsid', 'kb', 'pos', 'a1', 'a2']
print(lifted_bim)

# use only lifted snps
lifted_bim['rsid'].to_csv(f'{geno_path}_hg19_lifted.snplist', sep='\t', header=False, index=False)

plink_extract_cmd = f'\
plink --bfile {geno_path}\
 --extract {geno_path}_hg19_lifted.snplist\
 --make-bed\
 --out {geno_path}_hg19_lifted'

shell_do(plink_extract_cmd)


# move bim with old positions to new file
!mv {geno_path}_hg19_lifted.bim {geno_path}_hg19_lifted_old_positions.bim

# write lifted bim to _hg19_lifted genotype name
lifted_bim.to_csv(f'{geno_path}_hg19_lifted.bim', sep='\t', header=False, index=False)

#check number of IDs in original .bim file
!cat {datadir}/qc_genotypes_twas_hg19_lifted.bim | wc -l

#look at the bim file (variance, chromosome and position, allele1 and allele2)
! head {datadir}/qc_genotypes_twas_hg19_lifted.bim

Liftover control genotype from hg38 to hg19

# set paths for controls
control_geno_path = f'{datadir}/qc_genotypes_twas_control'
pheno_path = f'{datadir}/twas_expression_control.txt'
covar_path = f'{datadir}/covariates_control.txt'

# get chrN:start-end positions for liftover of genotype from hg38 to hg19 to match 1kG LD ref
lift_outname_cont = f'{datadir}/control_geno_hg38_positions.bed'
bim_cont = pd.read_csv(f'{geno_path}_control.bim', sep='\t', header=None)
bim_cont.columns = ['chr', 'rsid', 'kb', 'pos', 'a1', 'a2']
bim_cont['chr'] = 'chr' + bim_cont['chr'].astype('str')
bim_cont['end'] = bim_cont['pos'] + 1
lift_out_cont = bim_cont[['chr', 'pos', 'end', 'rsid' ]].copy()
lift_out_cont.to_csv(lift_outname_cont, sep='\t', header=False, index=False)

# create command for liftover
liftOver = f'{basedir}/liftover/liftOver'
chainfile = f'{basedir}/liftover/hg38ToHg19.over.chain.gz'
liftover_output_cont = f'{basedir}/control_geno_hg19_positions.bed'
unlifted_cont = f'{basedir}/control_unlifted.bed'

liftover_cont_cmd = f'\
{liftOver} {lift_outname_cont} {chainfile} {liftover_output_cont} {unlifted_cont}'

shell_do(liftover_cont_cmd)
liftover_cont_cmd

# create map update file
lifted_cont = pd.read_csv(liftover_output_cont, sep='\t', header=None, names=['chr', 'pos', 'end', 'rsid' ])

bim_cont = pd.read_csv(f'{geno_path}_control.bim', sep='\t', header=None, names=['chr', 'rsid', 'kb', 'pos', 'a1', 'a2'])
bim_lift_merge_cont = bim_cont.merge(lifted_cont, how='right', on='rsid')
lifted_bim_cont = bim_lift_merge_cont[['chr_x', 'rsid', 'kb', 'pos_y', 'a1', 'a2']].copy()
lifted_bim_cont.columns = ['chr', 'rsid', 'kb', 'pos', 'a1', 'a2']

# use only lifted snps
lifted_bim_cont['rsid'].to_csv(f'{geno_path}_control_hg19_lifted.snplist', sep='\t', header=False, index=False)

plink_extract_cont_cmd = f'\
plink --bfile {geno_path}_control\
 --extract {geno_path}_control_hg19_lifted.snplist\
 --make-bed\
 --out {geno_path}_control_hg19_lifted'

shell_do(plink_extract_cont_cmd)

# move bim with old positions to new file
!mv {geno_path}_control_hg19_lifted.bim {geno_path}_control_hg19_lifted_old_positions.bim

# write lifted bim to _hg19_lifted genotype name
lifted_bim_cont.to_csv(f'{geno_path}_control_hg19_lifted.bim', sep='\t', header=False, index=False)

#check number of IDs in original .bim file
!cat {datadir}/qc_genotypes_twas_control_hg19_lifted.bim | wc -l

#look at the bim file (variance, chromosome and position, allele1 and allele2)
! head {datadir}/qc_genotypes_twas_control_hg19_lifted.bim
