==> T-WAS/Jupyter_notebook/3_calculate_weights_case.ipynb <==
#clone fusion_twas library before running this
!git clone https://github.com/gusevlab/fusion_twas.git

import subprocess
import sys
import os
import shutil
import pandas as pd

def shell_do(command, log=False, return_log=False):
    print(f'Executing: {(" ").join(command.split())}', file=sys.stderr)

    res=subprocess.run(command.split(), stdout=subprocess.PIPE)

    if log:
        print(res.stdout.decode('utf-8'))
    if return_log:
        return(res.stdout.decode('utf-8'))

#set paths
basedir = '/data/songy4/twas'
datadir = f'{basedir}/data_folder'
fusiondir = f'{basedir}/fusion_twas'
geno_path = f'{datadir}/qc_genotypes_twas'
gene_list_path = f'{datadir}/gene_list.txt'
pheno_path = f'{datadir}/twas_expression_case.txt'
coord_path = f'{datadir}/twas_coordinate.txt'
covar_path = f'{datadir}/covariates_case.txt'
gcta = f'{fusiondir}/gcta_nr_robust'
gemma = f'gemma-0.98.3-linux-static'
fusion_ldref_basename = f'{fusiondir}/LDREF/1000G.EUR'
fusion_compute_weights_script = f'{fusiondir}/FUSION.compute_weights.R'

!mkdir --parents output/weights
!mkdir --parents output/tmp


# Pipeline for TWAS run

# get gene list
##test with couple genes first before running all gene_list
gene_list_df = pd.read_csv(gene_list_path, sep='\t')
# gene_list = list(gene_list_df.ID)
gene_list = ['ENSG00000186092', 'ENSG00000187634', 'ENSG00000188976'] # for testing
# gene_list = ['ENSG00000186092']
pheno = pd.read_csv(pheno_path, sep='\t')
coords = pd.read_csv(coord_path, sep='\t')
covars = pd.read_csv(covar_path, sep='\t')

# now put together pipeline

compweights_swarmfile = f'{basedir}/compute_weights.swarm'

with open(compweights_swarmfile, 'w') as f:
    
    for gene in gene_list:
        OUT = f'output/tmp/{gene}'
        FINAL_OUT = f'output/weights/{gene}'
        #get chr start stop
        _chr = coords.loc[coords.ID == gene, 'X.Chr'].item()
        _start = coords.loc[coords.ID == gene, 'start'].item()-0.5e6
        _stop = coords.loc[coords.ID == gene, 'end'].item()+0.5e6

        if _start < 0:
            _start = 0
        
        _start = int(_start)
        _stop = int(_stop)
        
#         _temp_name = f'{gene}_temp'
#         _gene_temp = f'{tempdir}/{_temp_name}'
        
        # pheno per gene
#         _phenoname = f'{_gene_temp}.pheno'
        
        # write pheno file per gene
        pheno[['FID','IID', gene]].to_csv(f'{OUT}.pheno', sep='\t', header=False, index=False)
#         pheno[['FID','IID', gene]].to_csv(_phenoname, sep='\t', header=False, index=False)
        
        plink_cmd = f'\
plink --bfile {geno_path}_hg19_lifted \
--pheno {OUT}.pheno \
--keep {OUT}.pheno \
--chr {_chr} \
--from-bp {_start} \
--to-bp {_stop} \
--extract {fusion_ldref_basename}.{_chr}.bim \
--make-bed \
--out {OUT}'  

#        shell_do(plink_cmd, log=True, return_log=True)

        fusion_cmd = f'\
Rscript {fusion_compute_weights_script} \
--bfile {OUT} \
--tmp {OUT}.tmp \
--out {FINAL_OUT} \
--PATH_gemma {gemma} \
--PATH_plink plink \
--PATH_gcta {gcta} \
--covar {covar_path} \
--verbose 2 \
--save_hsq \
--models top1,lasso,enet'
##include "blup,bslmm" to the --models in fusion_cmd for future     
# 
#        shell_do(fusion_cmd, log=True, return_log=True)

        f.write(f'{plink_cmd} && {fusion_cmd}\n')
    f.close()

# run swarm
swarm_cmd = f'swarm -f {compweights_swarmfile} -g 16 -t 16 --time=10:00:00 --logdir swarm --gres=lscratch:20 --module plink,GEMMA/0.96 --partition=norm'
# shell_do(swarm_cmd)
print(swarm_cmd)
!{swarm_cmd}
##check swarm jobs state in hpc.nih.gov/dashboard/
##check .e, .o, and .log files of swarm jobs in swarm folder (same name as swarm jobid in hpc dashboard)

#check number of weights in output/weights folder 
###run this in terminal

#go to the weights folder
!cd /data/songy4/twas/output/weights
#check how many files are in total
! ls | wc -l
#there are 27509
#check how many .RDat file is in weights folder
!ls -lR ./*.RDat | wc -l
#there are 9534 RDat files

TWAS

#open gene_chr_start_stop.tab -- this has chr and position info
chr_pos_df = pd.read_csv(coord_path, sep='\t')
#rearrange the columns 
chr_pos_df = chr_pos_df[['ID', 'X.Chr','start','end']]
#change column names in chr_pos_df
chr_pos_df.columns = ['ID', 'CHR', 'P0', 'P1']

chr_pos_df

#make gene_path.csv
! ls -l ./output/weights | grep -E 'ENSG|.RDat' | awk '{print $9", ./weights/"$9}' | sed 's/.wgt.RDat,/,/g' > ./data_folder/gene_path.csv

## list files in weights folder, search lines with same pattern as 'ENSG'
## scan file line by line, splits each input line
## substitute .wgt.RDat, with , in all occurrences (/g means global replacement)

#import gene_path.csv 
path = pd.read_csv(f"{datadir}/gene_path.csv", names=['ID', 'WGT'])
path.head

#remove .hsq rows
path = path[~path.WGT.str.contains(".hsq")]
path.head

#save path as csv -- run only once for download
path.to_csv(r'./data_folder/gene_path_case.csv' ,index=False)

#merge chr_pos_df and path
merged = chr_pos_df.merge(path, how='inner', on='ID')
merged.head

#check chr 
chr_list = merged['CHR'].unique().tolist()
chr_list

#drop ./weights/ from WGT
merged['WGT'] = merged['WGT'].str.replace('./weights/', '')
merged.head

#rearrange columns and save the file as synapse.pos
output = merged[['WGT', 'ID', 'CHR', 'P1', 'P0']]
output.to_csv(r'./data_folder/synapse.pos', index=False, sep="\t")

#set paths
sumstat_path = f'{datadir}/meta.txt'
weights_path = f'{datadir}/synapse.pos'
weights_dir = f'{basedir}/output/weights/'
fusion_ldref_basename = f'{fusiondir}/LDREF/1000G.EUR'
fusion_assoc_script = f'{fusiondir}/FUSION.assoc_test.R'

!mkdir --parents output/pd

sumstat = pd.read_csv(sumstat_path, sep='\t')

##check if weights_dir should be f'{basedir}/output/weights'####

#fusion calculation -- take about 3 min each chromosome 
for i in range(1, 23):
    PD_OUT = f'output/pd/PD.{i}.dat'
    fusion_cmd_i = f'\
    Rscript {fusion_assoc_script} \
    --sumstats {sumstat_path} \
    --weights {weights_path} \
    --weights_dir {weights_dir} \
    --ref_ld_chr {fusion_ldref_basename}. \
    --chr {i} \
    --out {PD_OUT}'
    shell_do(fusion_cmd_i)

# running list of questions
1. what about non-autosome transcripts? i.e. MT
2. how many snps in common between ref and geno?
3. how many snps per transcript should we expect (0-500?)


p_in = 'tmp/ENSG00000188976.tmp.cv'
f"gemma -miss 1 -maf 0 -r2 1 -rpace 1000 -wpace 1000 -bfile {p_in} -bslmm 2 -o ENSG00000188976"
